Use Blaze_Import's parent directory for Version.h
Do not use `CMAKE_CURRENT_LIST_DIR` when reading
`blaze/system/Version.h` as this assumes that the `blaze` directory is
present next to the currently processing CMake file, i.e. the one that
calls `Blaze_Import()`. Instead, determine the parent directory of
`Blaze_Import.cmake`, which we know to be the root of the Blaze source,
and use it as the base path for reading `Version.h`.