Fix the include directory set on the Blaze target.
The call to `target_include_directories()` to set the include directory
on the `Blaze` `INTERFACE` target is using an undefined variable and
has an extra parameter, `GLOBAL`, that incorrectly sets an include
directory of `GLOBAL` on the target.

Rename `Blaze_Import_PARENT_DIR_PATH` to `Blaze_ROOT_DIR_PATH` since it
points to the project root and use correct the call to set the include
directory by using it.