Do not guard against multiple Blaze_Import() calls
Guarding against multiple `Blaze_Import()` calls will cause subsequent
invocations of CMake to improperly configure the `Blaze` target. To
avoid this we remove the guard inside of `Blaze_Import()` and always run
the function. There should be no adverse side effects and indeed,
because the `Blaze` target is an `INTERFACE` target it must be
reconfigured since its settings (include directories, for example) are
not persisted with the build system itself.

Resolves #422